requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
DAWG-Python==0.7.2
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.2
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
sniffio==1.3.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
SQLAlchemy~=2.0.38